atualizando models com long/ banco de dados atualizado/mapeamento atualizado

+++ Zoozinho/ZooConsole/Models/Animal.cs
@@ -8,19 +8,19 @@ namespace ZooConsole.Models
 {
     public class Animal
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nome { get; set; }
         public string Sexo { get; set; }
         public int Idade { get; set; }
-        public double Peso { get; set; }
+        public decimal Peso { get; set; }
 
-        public int EspecieId { get; set; }
+        public long EspecieId { get; set; }
         public Especie Especie { get; set; }
 
-        public int? HabitatId { get; set; }
+        public long? HabitatId { get; set; }
         public Habitat Habitat { get; set; }
 
-        public int? GalpaoId { get; set; }
+        public long? GalpaoId { get; set; }
         public Galpao Galpao { get; set; }
 
         public string Localizacao =>
