+++ Zoozinho/ZooConsole/Models/Especie.cs
@@ -8,11 +8,11 @@ namespace ZooConsole.Models
 {
     public class Especie
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nome { get; set; }
         public string Alimentacao { get; set; }
         public string Comportamento { get; set; }
-        public int CategoriaId { get; set; }
+        public long CategoriaId { get; set; }
         public Categoria Categoria { get; set; }
         public List<Animal> Animais { get; set; } = new List<Animal>();
     }
