+++ Zoozinho/ZooConsole/Models/Recinto.cs
@@ -8,9 +8,9 @@ namespace ZooConsole.Models
 {
     public class Recinto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nome { set; get; }
-        public int CategoriaId { get; set; }
+        public long CategoriaId { get; set; }
         public Categoria Categoria { get; set; }
         public int CapacidadeMaxHabitats { get; set; }
         public List<Habitat> Habitats { get; set; } = new List<Habitat>();
